ongoing notification: add preference option
