+++ app/src/main/java/com/eveningoutpost/dexdrip/UtilityModels/XdripNotificationCompat.java
@@ -32,27 +32,5 @@ public class XdripNotificationCompat extends NotificationCompat {
             return builder.build(); // standard pre-oreo behaviour
         }
     }
-
-    @TargetApi(Build.VERSION_CODES.O)
-    public static Notification build(Notification.Builder builder) {
-        if ((Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.O)) {
-            if (Pref.getBooleanDefaultFalse("use_notification_channels")) {
-                // get dynamic channel based on contents of the builder
-                try {
-                    final String id = NotificationChannels.getChan(builder).getId();
-                    builder.setChannelId(id);
-                } catch (NullPointerException e) {
-                    //noinspection ConstantConditions
-                    builder.setChannelId(null);
-                }
-            } else {
-                //noinspection ConstantConditions
-                builder.setChannelId(null);
-            }
-            return builder.build();
-        } else {
-            return builder.build(); // standard pre-oreo behaviour
-        }
-    }
 }
