+++ app/src/main/res/xml/pref_notifications.xml
@@ -82,6 +82,13 @@
                     android:summary="Use the Android 8+ notification channel feature"
                     android:title="Notification Channels" />
 
+                <CheckBoxPreference
+                    android:defaultValue="false"
+                    android:dependency="use_notification_channels"
+                    android:key="ongoing_notification_channel"
+                    android:summary="May trigger ambient display notifications on some devices. Needs testing!"
+                    android:title="Use ongoing notification channel" />
+
             </PreferenceScreen>
